requirements.txt
Flask
requests
Flask-SQLAlchemy
Flask-bcrypt
bleach
pyjwt